requirements.txt
Faker==19.11.0
graphviz==0.20.1
greenlet==3.0.0
neo4j==5.13.0
networkx==3.2.1
ply==3.11
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
